First pass at new client.
Based on lessons from last few attempts, I'm going a different route
this time. Instead of focusing on the user API first, I have focused on
the rate limiting. The in-process rate limiter looks fairly complete,
and is a couple of steps away from a rate-limit server that can
coordinate between multiple scrapers.