Add first naive implementation of champ select solver.
The solver uses a brute force approach to solving for the best ban and
pick strategy given a utility function. The goal is to maximize the
difference between your utiity and your opponent's utility.

There are a couple of optimizations to prune the search size by marking
certain states equivalent, but there is clearly much more work to be
done here to prune the tree to a reasonable size.